refactor(routes): extract play now route
